requirements.txt
discord.py
openai
flask
flask-cors
httpx
python-dotenv
pymongo
numpy
